+++ addons/br/nfse/nfse.go
@@ -26,7 +26,9 @@ func newAddon() *tax.AddonDef {
 			i18n.EN: "Brazil NFS-e 1.X",
 		},
 		Extensions: extensions,
+		Identities: identities,
 		Validator:  validate,
+		Normalizer: normalize,
 	}
 }
 
@@ -40,4 +42,11 @@ func validate(doc any) error {
 		return validateItem(obj)
 	}
 	return nil
+}
+
+func normalize(doc any) {
+	switch obj := doc.(type) {
+	case *bill.Invoice:
+		normalizeSupplier(obj.Supplier)
+	}
 }